refactor(provider): extract ProviderLoopConfig and ProviderLogger from ProviderLoop
First, low-risk step of breaking up the ProviderLoop god file: move the two
standalone top-level value types out of the 3.6k-line ProviderLoop.swift into
their own single-responsibility files (no actor internals touched, no access
loosened). The loop file now holds the actor + its behavior, not its inputs
or its logger type.

Note: the actor's instance-method sections are intentionally NOT split across
files here — ProviderLoop deliberately keeps method extensions in-file to reach
its `private` model registry without loosening ~40 members to `internal` (see
the documented rationale at the local-endpoint extension). That deeper split is
deferred pending an explicit encapsulation decision.

Behavior-preserving: swift build green; swift test 1052 pass / 0 fail.
